backend/spotify/requirements.txt
Django>=4.2.1,<4.3
psycopg2-binary>=2.9.6,<2.10
djangorestframework>=3.14.0
Pillow>=9.0.0